+++ mai-locations.php
@@ -4,7 +4,7 @@
  * Plugin Name:       Mai Locations
  * Plugin URI:        https://bizbudding.com
  * Description:       A custom post type with info/address/map fields to manage locations.
- * Version:           1.1.0
+ * Version:           2.0.0
  * Requires PHP:      8.3
  * Requires at least: 6.9
  * Author:            BizBudding
@@ -93,7 +93,7 @@ final class Mai_Locations_Plugin {
 	private function setup_constants() {
 		// Plugin version.
 		if ( ! defined( 'MAI_LOCATIONS_VERSION' ) ) {
-			define( 'MAI_LOCATIONS_VERSION', '1.1.0' );
+			define( 'MAI_LOCATIONS_VERSION', '2.0.0' );
 		}
 
 		// Plugin Folder Path.